libft

+++ 02_libft/srcs/ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: kemesure <kemesure@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/11 16:35:45 by kemesure          #+#    #+#             */
-/*   Updated: 2017/12/03 16:03:33 by kemesure         ###   ########.fr       */
+/*   Updated: 2017/12/06 18:30:47 by kemesure         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,5 +19,5 @@ int		ft_strcmp(const char *s1, const char *s2)
 	i = 0;
 	while (s1[i] == s2[i] && s1[i] && s2[i])
 		i++;
-	return (s1[i] - s2[i]);
+	return ((int)s1[i] - (int)s2[i]);
 }
